Fix a couple of Z linux compile warnings
git-svn-id: https://svn.code.sf.net/p/oorexx/code-0/main/trunk@4757 0b6cbdbe-3aab-466e-b73a-abd511dda0a2

+++ interpreter/classes/NumberStringClass.cpp
@@ -36,7 +36,7 @@
 /*                                                                            */
 /*----------------------------------------------------------------------------*/
 /******************************************************************************/
-/* REXX Kernel                                         NumberStringClass.c    */
+/* REXX Kernel                                                                */
 /*                                                                            */
 /* Primitive NumberString Class                                               */
 /*                                                                            */
@@ -3278,7 +3278,7 @@ RexxNumberString *RexxNumberString::newInstanceFromDouble(double number, size_t 
     char doubleStr[30];
     /* get double as a string value.     */
     /* Use digits as precision.          */
-    sprintf(doubleStr, "%.*g", precision + 2, number);
+    sprintf(doubleStr, "%.*g", (int)(precision + 2), number);
     resultLen = strlen(doubleStr);       /* Compute length of floatString     */
                                          /* Create new NumberString           */
     result = new (resultLen) RexxNumberString (resultLen, precision);
